requirements.txt
ansicon==1.89.0
asgiref==3.5.2
billiard==3.6.4.0
blinker==1.4
cached-property==1.5.2
cement==2.8.2
certifi==2021.5.30
charset-normalizer==2.1.0
dj-database-url==0.5.0
Django==3.2.4
django-cors-headers==3.7.0
django-crontab==0.7.1
django-debug-toolbar==3.2.2
django-email-verification==0.1.0
django-password-validators==1.6.1
django-rest-passwordreset==1.2.1
djangorestframework==3.12.4
djangorestframework-api-key==2.0.0
docopt==0.6.2
future==0.16.0
gprof2dot==2021.2.21
idna==3.3
pathspec==0.5.9
Pillow==8.4.0
prompt-toolkit==3.0.20
psycopg2==2.8.6
pycodestyle==2.7.0
pycparser==2.20
PyGetWindow==0.0.9
Pygments==2.10.0
PyJWT==1.7.1
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.1.4
pyrsistent==0.18.0
PyScreeze==0.1.27
python-decouple==3.5
python3-xlib==0.15
PyTweening==1.0.3
pytz==2021.1
PyYAML==5.4.1
razorpay==1.2.0
requests==2.28.1
semantic-version==2.8.5
six==1.14.0
soupsieve==2.2.1
sqlparse==0.4.2
termcolor==1.1.0
texttable==1.6.4
toml==0.10.2
twilio==6.63.2
tzlocal==2.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.1.9
whitenoise==5.2.0
wrapt==1.14.1
